DSTEG-11: Refactoring.
