+++ dst_entity_generate.services.yml
@@ -14,6 +14,4 @@ services:
       - '@keyvalue'
       - '@config.factory'
       - '@entity_type.manager'
-  dst_entity_generate.helper:
-    class: Drupal\dst_entity_generate\Services\Helper
-    arguments: ['@config.factory', '@module_handler', '@entity_display.repository', '@entity_type.manager', '@dst_entity_generate.google_sheet_api', '@logger.factory', '@entity_display.repository']
+      - '@entity_display.repository'
